多个数据源

+++ fisk-factory-access/fisk-factory-access-entity/src/main/java/com/fisk/dataaccess/dto/datasource/DataSourceInfoDTO.java
@@ -10,6 +10,6 @@ public class DataSourceInfoDTO {
 
     public String name;
 
-    public Integer id;
+    public long id;
 
 }
